% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simu_data}
\alias{simu_data}
\title{Simulated data for SHIR demo}
\format{
A data list with 5 elements:
\describe{
\item{X}{...}
\item{Y}{...}
\item{beta}{...}
\item{mu}{...}
\item{alpha}{...}
}
}
\usage{
simu_data
}
\description{
Simulated data for SHIR demo
}
\keyword{datasets}
