% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_fit.R
\name{Local_fit}
\alias{Local_fit}
\title{Local_fit: Fit lasso and derive summary statistics at each local site.}
\usage{
Local_fit(Y, X, lambda_lst = NULL)
}
\arguments{
\item{Y}{individual response vector}

\item{X}{individual covariates matrices}

\item{lambda_lst}{candidate set of tuning parameters, NULL for choose the default range in glmnet}
}
\value{
hessian: derived Hessian matrix

gradient: derived gradient vector

beta: local lasso estimator
}
\description{
Local_fit: Fit lasso and derive summary statistics at each local site.
}
