% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHIR_fit.R
\name{SHIR_fit}
\alias{SHIR_fit}
\title{SHIR_fit: aggregate the locally derived summary data using the proposed method in SHIR paper.}
\usage{
SHIR_fit(H_lst, d_lst, n_lst, lambda_lst, lambda_g_lst, tune = "BIC")
}
\arguments{
\item{H_lst}{list of locally derived Hessian matrix}

\item{d_lst}{list of locally derived gradient vector}

\item{n_lst}{vector of the sample sizes at local sites}

\item{lambda_lst}{candidate set of the tuning parameter for mu,
(corresponds to sqrt(N)lambda in the SHIR paper)}

\item{lambda_g_lst}{candidate set of the tuning parameter for alpha,
(corresponds to sqrt(N)*lambdalambda_g in the SHIR paper)}

\item{tune}{the information criterion used for model selection:
options include 'AIC', 'BIC' and 'RIC', which put different weights on the degree of freedom.}
}
\value{
min.lambda: the selected tuning parameters with the minimum information criterion

min.beta: matrices of the fitted SHIR estimator
(the m-th column loads the fitted coefficient (beta) for the m-th site,
the first row is the fitted intercepts)
}
\description{
SHIR_fit: aggregate the locally derived summary data using the proposed method in SHIR paper.
}
