% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simu_data}
\alias{simu_data}
\title{Simulated data for the example.}
\format{
A data list with 5 elements:
\describe{
\item{X}{list of individual covariates matrices of the local sites}
\item{Y}{list of individual response vectors of the local sites}
\item{beta}{list of true coefficients (beta) of the local sites}
\item{mu}{vector of the true mu (homogeneous effect)}
\item{alpha}{list of true alpha (heterogeneous effect) of the local sites}
}
}
\usage{
simu_data
}
\description{
Simulated data for the example.
}
\keyword{datasets}
